+++ src/backend/joanie/core/enums.py
@@ -97,3 +97,10 @@ ENROLLMENT_STATE_CHOICES = (
 CONTRACT_DEFINITION = "contract_definition"
 
 CONTRACT_NAME_CHOICES = ((CONTRACT_DEFINITION, _("Contract Definition")),)  # default
+
+# For certification names choices
+DEGREE = "degree"
+
+CERTIFICATE_NAME_CHOICES = (
+    (DEGREE, _("Degree")),
+)